requirements.txt
flask==3.0.0
spacy==3.7.2
newsapi-python==0.2.7
python-dotenv==1.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl